+++ core/functions/reply_markup.py
@@ -4,8 +4,8 @@ from core.commands import ADMIN_COMMAND_ORDER, ADMIN_COMMAND_STATUS, ADMIN_COMMA
     ADMIN_COMMAND_FIRE_UP, ADMIN_COMMAND_SQUAD_LIST, USER_COMMAND_ME, USER_COMMAND_TOP, USER_COMMAND_SQUAD, \
     USER_COMMAND_STATISTICS, USER_COMMAND_BUILD, USER_COMMAND_CONTACTS, ADMIN_COMMAND_ADMINPANEL, \
     USER_COMMAND_SQUAD_REQUEST, USER_COMMAND_BACK, TOP_COMMAND_ATTACK, TOP_COMMAND_DEFENCE, TOP_COMMAND_EXP, \
-    STATISTICS_COMMAND_EXP, USER_COMMAND_SQUAD_LEAVE, ADMIN_COMMAND_REPORTS, TOP_COMMAND_GLOBAL_BUILD, \
-    TOP_COMMAND_WEEK_BUILD, TOP_COMMAND_WEEK_BATTLES
+    STATISTICS_COMMAND_EXP, USER_COMMAND_SQUAD_LEAVE, ADMIN_COMMAND_REPORTS, TOP_COMMAND_BUILD, \
+    TOP_COMMAND_BATTLES
 
 
 def generate_admin_markup(full=False):
@@ -28,8 +28,8 @@ def generate_user_markup(is_admin=False):
 
 def generate_top_markup():
     buttons = [[KeyboardButton(TOP_COMMAND_ATTACK), KeyboardButton(TOP_COMMAND_DEFENCE),
-                KeyboardButton(TOP_COMMAND_EXP), KeyboardButton(TOP_COMMAND_GLOBAL_BUILD)],
-               [KeyboardButton(TOP_COMMAND_WEEK_BUILD), KeyboardButton(TOP_COMMAND_WEEK_BATTLES)],
+                KeyboardButton(TOP_COMMAND_EXP), KeyboardButton(TOP_COMMAND_BUILD),
+                KeyboardButton(TOP_COMMAND_BATTLES)],
                [KeyboardButton(USER_COMMAND_BACK)]]
     return ReplyKeyboardMarkup(buttons, True)
