+++ core/functions/top.py
@@ -1,15 +1,18 @@
 from sqlalchemy import func, text as text_, tuple_
 from telegram import Update, Bot
 
+from core.functions.inline_markup import generate_build_top, generate_battle_top
 from core.functions.reply_markup import generate_top_markup
 from core.texts import MSG_TOP_ABOUT, MSG_TOP_FORMAT, MSG_TOP_ATTACK, MSG_TOP_DEFENCE, MSG_TOP_EXPERIENCE, \
-    MSG_TOP_GLOBAL_BUILDERS, MSG_TOP_WEEK_BUILDERS, MSG_TOP_WEEK_WARRIORS
-from core.types import user_allowed, Character, BuildReport, Report
+    MSG_TOP_GLOBAL_BUILDERS, MSG_TOP_WEEK_BUILDERS, MSG_TOP_WEEK_WARRIORS, MSG_SQUAD_TOP_FORMAT
+from core.types import user_allowed, Character, BuildReport, Report, Squad, SquadMember
 from core.utils import send_async
 
 from config import CASTLE
 
 from datetime import datetime, timedelta
+
+TOP_START_DATE = datetime(2017, 11, 11)
 
 
 @user_allowed
@@ -78,6 +81,42 @@ def exp_top(bot: Bot, update: Update, session):
                text=text)
 
 
+def gen_top_msg(data, user_id, header, icon):
+    text = header
+    str_format = MSG_TOP_FORMAT
+    for i in range(min(10, len(data))):
+        text += str_format.format(i + 1, data[i][0].name, data[i][0].level,
+                                  data[i][1], icon)
+    if len(data) and hasattr(data[0][0], 'user_id') and user_id in [build[0].user_id for build in data]:
+        if user_id not in [build[0].user_id for build in data[:10]]:
+            for i in range(10, len(data)):
+                if data[i][0].user_id == user_id:
+                    text += '...\n'
+                    text += str_format.format(i, data[i - 1][0].name, data[i - 1][0].level,
+                                              data[i - 1][1], icon)
+                    text += str_format.format(i + 1, data[i][0].name, data[i][0].level,
+                                              data[i][1], icon)
+                    if i != len(data) - 1:
+                        text += str_format.format(i + 2, data[i + 1][0].name, data[i + 1][0].level,
+                                                  data[i + 1][1], icon)
+                    break
+    return text
+
+
+def gen_squad_top_msg(data, counts, header, icon):
+    text = header
+    str_format = MSG_SQUAD_TOP_FORMAT
+    for i in range(min(10, len(data))):
+        squad_count = 1
+        for squad, count in counts:
+            if squad.chat_id == data[i][0].chat_id:
+                squad_count = count
+                break
+        text += str_format.format(i + 1, data[i][0].squad_name, squad_count,
+                                  data[i][1], icon, round(data[i][1]/squad_count, 2), icon)
+    return text
+
+
 @user_allowed
 def global_build_top(bot: Bot, update: Update, session):
     actual_profiles = session.query(Character.user_id, func.max(Character.date)). \
@@ -87,33 +126,26 @@ def global_build_top(bot: Bot, update: Update, session):
         .filter(tuple_(Character.user_id, Character.date)
                 .in_([(a[0], a[1]) for a in actual_profiles]),
                 Character.date > datetime.now() - timedelta(days=7))\
-        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id).group_by(Character) \
-        .filter(BuildReport.date > datetime(2017, 12, 1))\
+        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id) \
+        .filter(BuildReport.date > TOP_START_DATE).group_by(Character)\
         .order_by(func.count(BuildReport.user_id).desc())
     if CASTLE:
         builds = builds.filter(Character.castle == CASTLE)
     builds = builds.all()
-    text = MSG_TOP_GLOBAL_BUILDERS
-    str_format = MSG_TOP_FORMAT
-    for i in range(min(10, len(builds))):
-        text += str_format.format(i + 1, builds[i][0].name, builds[i][0].level,
-                                  builds[i][1], '⚒')
-    if update.message.from_user.id in [build[0].user_id for build in builds]:
-        if update.message.from_user.id not in [build[0].user_id for build in builds[:10]]:
-            for i in range(10, len(builds)):
-                if builds[i][0].user_id == update.message.from_user.id:
-                    text += '...\n'
-                    text += str_format.format(i, builds[i - 1][0].name, builds[i-1][0].level,
-                                              builds[i - 1][1], '⚒')
-                    text += str_format.format(i + 1, builds[i][0].name, builds[i][0].level,
-                                              builds[i][1], '⚒')
-                    if i != len(builds) - 1:
-                        text += str_format.format(i + 2, builds[i + 1][0].name, builds[i+1][0].level,
-                                                  builds[i + 1][1], '⚒')
-                    break
-    send_async(bot,
-               chat_id=update.message.chat.id,
-               text=text)
+    user_id = 0
+    if update.message:
+        user_id = update.message.from_user.id
+    elif update.callback_query:
+        user_id = update.callback_query.message.from_user.id
+    text = gen_top_msg(builds, user_id, MSG_TOP_GLOBAL_BUILDERS, '⚒')
+    markup = generate_build_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
 
 
 @user_allowed
@@ -126,33 +158,83 @@ def week_build_top(bot: Bot, update: Update, session):
         .filter(tuple_(Character.user_id, Character.date)
                 .in_([(a[0], a[1]) for a in actual_profiles]),
                 Character.date > datetime.now() - timedelta(days=7))\
-        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id).group_by(Character) \
-        .filter(BuildReport.date > today - timedelta(days=today.weekday()))\
+        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id) \
+        .filter(BuildReport.date > today - timedelta(days=today.weekday())).group_by(Character)\
         .order_by(func.count(BuildReport.user_id).desc())
     if CASTLE:
         builds = builds.filter(Character.castle == CASTLE)
     builds = builds.all()
-    text = MSG_TOP_WEEK_BUILDERS
-    str_format = MSG_TOP_FORMAT
-    for i in range(min(10, len(builds))):
-        text += str_format.format(i + 1, builds[i][0].name, builds[i][0].level,
-                                  builds[i][1], '⚒')
-    if update.message.from_user.id in [build[0].user_id for build in builds]:
-        if update.message.from_user.id not in [build[0].user_id for build in builds[:10]]:
-            for i in range(10, len(builds)):
-                if builds[i][0].user_id == update.message.from_user.id:
-                    text += '...\n'
-                    text += str_format.format(i, builds[i - 1][0].name, builds[i-1][0].level,
-                                              builds[i - 1][1], '⚒')
-                    text += str_format.format(i + 1, builds[i][0].name, builds[i][0].level,
-                                              builds[i][1], '⚒')
-                    if i != len(builds) - 1:
-                        text += str_format.format(i + 2, builds[i + 1][0].name, builds[i+1][0].level,
-                                                  builds[i + 1][1], '⚒')
-                    break
-    send_async(bot,
-               chat_id=update.message.chat.id,
-               text=text)
+    user_id = 0
+    if update.message:
+        user_id = update.message.from_user.id
+    elif update.callback_query:
+        user_id = update.callback_query.message.from_user.id
+    text = gen_top_msg(builds, user_id, MSG_TOP_WEEK_BUILDERS, '⚒')
+    markup = generate_build_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
+
+
+@user_allowed
+def week_squad_build_top(bot: Bot, update: Update, session):
+    actual_profiles = session.query(Character.user_id, func.max(Character.date)). \
+        group_by(Character.user_id)
+    actual_profiles = actual_profiles.all()
+    today = datetime.today()
+    builds = session.query(Squad, func.count(BuildReport.user_id))\
+        .join(SquadMember).join(Character, SquadMember.user_id == Character.user_id)\
+        .filter(tuple_(Character.user_id, Character.date)
+                .in_([(a[0], a[1]) for a in actual_profiles]),
+                Character.date > datetime.now() - timedelta(days=7))\
+        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id) \
+        .filter(BuildReport.date > today - timedelta(days=today.weekday()))\
+        .order_by(func.count(BuildReport.user_id).desc())
+    if CASTLE:
+        builds = builds.filter(Character.castle == CASTLE)
+    builds = builds.group_by(Squad).all()
+    counts = session.query(Squad, func.count(SquadMember.user_id)).join(SquadMember).group_by(Squad)
+    text = gen_squad_top_msg(builds, counts, MSG_TOP_WEEK_BUILDERS, '⚒')
+    markup = generate_build_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
+
+
+@user_allowed
+def global_squad_build_top(bot: Bot, update: Update, session):
+    actual_profiles = session.query(Character.user_id, func.max(Character.date)). \
+        group_by(Character.user_id)
+    actual_profiles = actual_profiles.all()
+    builds = session.query(Squad, func.count(BuildReport.user_id))\
+        .join(SquadMember).join(Character, SquadMember.user_id == Character.user_id)\
+        .filter(tuple_(Character.user_id, Character.date)
+                .in_([(a[0], a[1]) for a in actual_profiles]),
+                Character.date > datetime.now() - timedelta(days=7))\
+        .outerjoin(BuildReport, BuildReport.user_id == Character.user_id) \
+        .filter(BuildReport.date > TOP_START_DATE) \
+        .order_by(func.count(BuildReport.user_id).desc())
+    if CASTLE:
+        builds = builds.filter(Character.castle == CASTLE)
+    builds = builds.group_by(Squad).all()
+    counts = session.query(Squad, func.count(SquadMember.user_id)).join(SquadMember).group_by(Squad)
+    text = gen_squad_top_msg(builds, counts, MSG_TOP_GLOBAL_BUILDERS, '⚒')
+    markup = generate_build_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
 
 
 @user_allowed
@@ -165,32 +247,56 @@ def week_battle_top(bot: Bot, update: Update, session):
         .filter(tuple_(Character.user_id, Character.date)
                 .in_([(a[0], a[1]) for a in actual_profiles]),
                 Character.date > datetime.now() - timedelta(days=7))\
-        .outerjoin(Report, Report.user_id == Character.user_id).group_by(Character) \
+        .outerjoin(Report, Report.user_id == Character.user_id) \
         .filter(Report.date > today - timedelta(days=today.weekday())) \
-        .filter(Report.earned_exp > 1) \
+        .filter(Report.earned_exp > 1).group_by(Character) \
         .order_by(func.count(Report.user_id).desc())
     if CASTLE:
         battles = battles.filter(Character.castle == CASTLE)
     battles = battles.all()
-    text = MSG_TOP_WEEK_WARRIORS
-    str_format = MSG_TOP_FORMAT
-    for i in range(min(10, len(battles))):
-        text += str_format.format(i + 1, battles[i][0].name, battles[i][0].level,
-                                  battles[i][1], '⚔')
-    if update.message.from_user.id in [build[0].user_id for build in battles]:
-        if update.message.from_user.id not in [build[0].user_id for build in battles[:10]]:
-            for i in range(10, len(battles)):
-                if battles[i][0].user_id == update.message.from_user.id:
-                    text += '...\n'
-                    text += str_format.format(i, battles[i - 1][0].name, battles[i-1][0].level,
-                                              battles[i - 1][1], '⚔')
-                    text += str_format.format(i + 1, battles[i][0].name, battles[i][0].level,
-                                              battles[i][1], '⚔')
-                    if i != len(battles) - 1:
-                        text += str_format.format(i + 2, battles[i + 1][0].name, battles[i+1][0].level,
-                                                  battles[i + 1][1], '⚔')
-                    break
-    send_async(bot,
-               chat_id=update.message.chat.id,
-               text=text)
+    user_id = 0
+    if update.message:
+        user_id = update.message.from_user.id
+    elif update.callback_query:
+        user_id = update.callback_query.message.from_user.id
+    text = gen_top_msg(battles, user_id, MSG_TOP_WEEK_WARRIORS, '⚔🛡')
+    markup = generate_battle_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
 
+
+@user_allowed
+def global_battle_top(bot: Bot, update: Update, session):
+    actual_profiles = session.query(Character.user_id, func.max(Character.date)). \
+        group_by(Character.user_id)
+    actual_profiles = actual_profiles.all()
+    battles = session.query(Character, func.count(Report.user_id))\
+        .filter(tuple_(Character.user_id, Character.date)
+                .in_([(a[0], a[1]) for a in actual_profiles]),
+                Character.date > datetime.now() - timedelta(days=7))\
+        .outerjoin(Report, Report.user_id == Character.user_id) \
+        .filter(Report.earned_exp > 1) \
+        .filter(Report.date > TOP_START_DATE).group_by(Character) \
+        .order_by(func.count(Report.user_id).desc())
+    if CASTLE:
+        battles = battles.filter(Character.castle == CASTLE)
+    battles = battles.all()
+    user_id = 0
+    if update.message:
+        user_id = update.message.from_user.id
+    elif update.callback_query:
+        user_id = update.callback_query.message.from_user.id
+    text = gen_top_msg(battles, user_id, MSG_TOP_WEEK_WARRIORS, '⚔🛡')
+    markup = generate_battle_top()
+    if update.message:
+        send_async(bot,
+                   chat_id=update.message.chat.id,
+                   text=text, reply_markup=markup)
+    elif update.callback_query:
+        bot.editMessageText(text, update.callback_query.message.chat.id, update.callback_query.message.message_id,
+                            reply_markup=markup)
