+++ main.py
@@ -21,8 +21,8 @@ from core.commands import ADMIN_COMMAND_STATUS, ADMIN_COMMAND_RECRUIT, ADMIN_COM
     ADMIN_COMMAND_GROUPS, ADMIN_COMMAND_FIRE_UP, USER_COMMAND_ME, USER_COMMAND_BUILD, USER_COMMAND_CONTACTS, \
     USER_COMMAND_SQUAD, USER_COMMAND_STATISTICS, USER_COMMAND_TOP, USER_COMMAND_SQUAD_REQUEST, USER_COMMAND_BACK, \
     TOP_COMMAND_ATTACK, TOP_COMMAND_DEFENCE, TOP_COMMAND_EXP, STATISTICS_COMMAND_EXP, USER_COMMAND_SQUAD_LEAVE, \
-    ADMIN_COMMAND_REPORTS, ADMIN_COMMAND_ADMINPANEL, TOP_COMMAND_GLOBAL_BUILD, TOP_COMMAND_WEEK_BUILD, \
-    TOP_COMMAND_WEEK_BATTLES
+    ADMIN_COMMAND_REPORTS, ADMIN_COMMAND_ADMINPANEL, TOP_COMMAND_BUILD, \
+    TOP_COMMAND_BATTLES
 from core.functions.activity import (
     day_activity, week_activity, battle_activity
 )
@@ -41,8 +41,9 @@ from core.functions.common import (
     delete_msg, delete_user,
     user_panel, web_auth)
 from core.functions.inline_keyboard_handling import (
-    callback_query, send_status, send_order, QueryType
+    callback_query, send_status, send_order
 )
+from core.functions.inline_markup import QueryType
 from core.functions.order_groups import group_list, add_group
 from core.functions.pin import pin, not_pin_all, pin_all, silent_pin
 from core.functions.profile import char_update, char_show, find_by_username, report_received, build_report_received, \
@@ -228,11 +229,9 @@ def manage_all(bot: Bot, update: Update, session, chat_data, job_queue):
                 def_top(bot, update)
             elif text == TOP_COMMAND_EXP.lower():
                 exp_top(bot, update)
-            elif text == TOP_COMMAND_GLOBAL_BUILD.lower():
-                global_build_top(bot, update)
-            elif text == TOP_COMMAND_WEEK_BUILD.lower():
+            elif text == TOP_COMMAND_BUILD.lower():
                 week_build_top(bot, update)
-            elif text == TOP_COMMAND_WEEK_BATTLES.lower():
+            elif text == TOP_COMMAND_BATTLES.lower():
                 week_battle_top(bot, update)
             elif text == USER_COMMAND_BUILD.lower():
                 send_async(bot,
